+++ php/src/Application/Command/VariableType/QueryingVisitor.php
@@ -240,6 +240,37 @@ class QueryingVisitor extends NodeVisitorAbstract
                     // the actual class is, so it's useless at the moment.
                 }
             }
+        } elseif ($node instanceof Node\Expr\FuncCall) {
+            if ($node->name instanceof Node\Name) {
+                $variableHandlingFunctionTypeMap = [
+                    'is_array'    => ['array'],
+                    'is_bool'     => ['bool'],
+                    'is_callable' => ['callable'],
+                    'is_double'   => ['float'],
+                    'is_float'    => ['float'],
+                    'is_int'      => ['int'],
+                    'is_integer'  => ['int'],
+                    'is_long'     => ['int'],
+                    'is_null'     => ['null'],
+                    'is_numeric'  => ['int', 'float', 'string'],
+                    'is_object'   => ['object'],
+                    'is_real'     => ['float'],
+                    'is_resource' => ['resource'],
+                    'is_scalar'   => ['int', 'float', 'string', 'bool'],
+                    'is_string'   => ['string']
+                ];
+
+                if (isset($variableHandlingFunctionTypeMap[$node->name->toString()])) {
+                    if (
+                        !empty($node->args) &&
+                        !$node->args[0]->unpack &&
+                        $node->args[0]->value instanceof Node\Expr\Variable &&
+                        $node->args[0]->value->name === $this->name
+                    ) {
+                        $types = $variableHandlingFunctionTypeMap[$node->name->toString()];
+                    }
+                }
+            }
         }
 
         return $types;